+++ src/router/iptools/db/examples/cxx/TpcbExample.cpp
@@ -1,7 +1,7 @@
 /*-
  * See the file LICENSE for redistribution information.
  *
- * Copyright (c) 1997, 2017 Oracle and/or its affiliates.  All rights reserved.
+ * Copyright (c) 1997, 2013 Oracle and/or its affiliates.  All rights reserved.
  *
  * $Id$
  */
@@ -246,7 +246,7 @@ usage()
 }
 
 TpcbExample::TpcbExample(const char *home, int cachesize, int flags)
-:	DbEnv((u_int32_t)0)
+:	DbEnv(0)
 {
 	u_int32_t local_flags;
