+++ src/router/iptools/db/test/cxx/TestLogc.cpp
@@ -1,7 +1,7 @@
 /*-
  * See the file LICENSE for redistribution information.
  *
- * Copyright (c) 2000, 2017 Oracle and/or its affiliates.  All rights reserved.
+ * Copyright (c) 2000, 2013 Oracle and/or its affiliates.  All rights reserved.
  *
  * $Id$
  */
@@ -32,7 +32,7 @@ static void show_dbt(ostream &os, Dbt *dbt)
 int main(int argc, char *argv[])
 {
 	try {
-		DbEnv *env = new DbEnv((u_int32_t)0);
+		DbEnv *env = new DbEnv(0);
 		DbTxn *dbtxn;
 		env->open(".", DB_CREATE | DB_INIT_LOG | 
 			  DB_INIT_TXN | DB_INIT_MPOOL, 0);
